requirements.txt
Flask
Keras
tensorflow==1.15
numpy
opencv-python
gitpython
pandas
pillow
